Added more clases, but nothing is really wired up
The model needs improving, currently LoanType leads nowhere, and the
amount of reward per unit of product type is stored in the enum class
FinancialProductType and also in the FinancialProduct's pointsPerAmount
attribute. I have to make a decision on how to model this and modify the
database accordingly.

Also, need to create to and fro, DTO's the services should not be
returning model entities to the controllers.