Added some TODOs
TODO 01: Modify this approach, the points per amount depend directly on the
product type. Most likely, the FinancialProductType will have to be
a full class instead of a enum, and have its own table.

TODO 02: Any idea how to retrieve the managed employees and the sales from
the role attribute given that it can be of subclass Manager or Seller?
NOTE: Seller is not implemented yet, it has to replace SalesPerson,
or I can transform SalesPerson into seller.