Prototype of the Employee Hierarchy
For now, I am storing both the list of sales and the list of managed
employees inside the Employee class. This is not ideal, as a manager
type employee does not have sales, and a seller type employee does not
have have managed employees.

Now, creating a sub-class for each does not seem flexible, because many
more roles might be added, and so, as sub-class for each does not seem
right. But, putting both lists in the same class seems even more
unacceptable. So, creating a sub-class for each role, and keeping the
role name (or id) inside the employee class, seems the best approach for
now.