Still working in persistence for the Employee tree
Modified the Manager and Seller classes to inherit from Employee, and
renamed the old classes withe suffix Orig. The new classes are not
mapped to new tables in the database, Employee is still the only table.
But now, sales and the list of managed employees is where it should be.
A better approach would have been some kind of composition, research on
this later.

Also, corrected all error in the dump.sql file, now the docker container
is created smoothly.