Improved version, now relying more on Spring Data Rest
For common CRUD operations, it seems that Spring Data Rest default
endpoints are quite enough, complex endpoints will have to be
personalized tho.

There's also the issue with the Data Rest not returning the objects id
as a field in the JSON response. But, we get the same functionality if
we extract, or just use, from the URLs in the '_links' object of the
response json. Just that, that would take some getting used to.