Have a bug in the customers REST functionality
Spring Data Rest generates endpoints for CRUD operations with paging and
sorting automatically by just extending from JpaRepository, but when you
extend from PagingAndSortingRepository, it seems things do not work so
smoothly.

Well is either that or, one can not have both active at the same time.

The customers endpoint is working for now, but further testing is
needed.