Spring Security with Jwt prototype working
Have not tested, or implemented the controllers for, signup and signin.
Much testing to do yet, but the application builds just fine.

Spring Security blocks all endpoints by default as soon as is included
as dependency. So, I had to disable it by setting to allow to /** (all
paths) in the configuration file. This setting has to be removed when
the testing is over.

Tried disabling Spring Security from the application.properties file,
but it seems that once a new configuration is set up using the
@Configuration annotation, it no longer takes into accou the properties
file.

This line (application.properties) is needed as a development secret key:
    token.signing.key=c2VjdXJlamF2YXNwcmluZ2Jvb3RrZXk=