Fixed SpringBoot missing test configuration
The Spring Boot app is already set to autoconfigure many features, but
tetst code needs to find the application file at the root of the whole
project (com.mybank.SpringbootdemoApplication.java). The problem was
that I hadn't renamed the package for the test class (only test class),
and if not found in the current package, it traverses upwards and tries
to find it, but, because the test code was under com.demo and the
application file was under com.mybank, it was never going to find it.
That is why Maven was throwing errors.