Succesfully dokerized the project, not perfectly tho
The project requires, for some reason, the database to be running to
even compile/build to a jar file. That being said, the server runs just
fine, and its already packaged with everything it needs to work.

Additionaly, had to change the application.properties file for the
database to be located at db instead of localhost, have to modify some
config in Eclipse or else it is not going to work there.